fix: better folder structure and erpnext settings
